addons/ps5_time_management/requirements.txt
paho-mqtt==2.0.0
flask==3.0.0
flask-cors==4.0.0

